Some improvements and observations
Next time, add a data field to the EmptyPane class regarding the position occupied by the EmptyPane, and set it during the instantiation of each and every instance. In that way you don't have to invoke the getPieceHolderNode function which sometimes results in misbehavior and directly retrieve the EmptyPane node through the pieceHolder using its position.