Resolved a vital issue of move function
Now we can move to a piece that was occupied by a piece before. Apparently move function seems to be working fine. However, when a pawn is about to en-passant the enemy one, showMoveables is not addressed correctly for some reason. Focus on the take function and implement the move for the king.