Additional implementation provided
After some quick testing I have observed that we can simply move the king to adjacent squares. Also we can castle but the rook doesn't move after castling. Furthermore, once we make a move with the king the turn is not transferred to the opponent. Thus, we weren't able to take a piece which was not guarded by any other piece using the king.