Fixes and additional observation
Also after moving a king to middle squares I have observed that it cannot move to squares which are reachable by enemy pawns but are not diagonal to them. We must fix this issue as well. I have resolved the issue regarding bishop not checking the king. Nevertheless, further testing is required.