Major observations
We cannot move a piece other than king to cover our king when the king is checked. It's very crucial that you resolve that. Also I observed occasionally we may not be able to rook for some reason temporarily. In addition, we are able to check the enemy king by moving our pawn up to an empty pane, but we cannot check it when we take it regularly or by en-passant.