Observations
The problem of not being able to cover a checked king with a friendly piece which can obstruct the path might be correlated to some problem identical to what we had on isGuarded function. Check the isChecked function of the app. Apparently, the possible state we pass is represented correctly. I tried to alternate the static showMoveables in corresponding piece classes such that they exclude occupied squares but it didn't work and I reverted back.