Observations & Major Fix
I have observed that during the isChecked function's execution, when we retrieve the bishop's moveables we don't incorporate the enemy king pos, that is why we are able to move a friendly piece which ultimately exposes our king. I have added a very vital and crucial logic element to bishop's static showMoveables. Now we cannot expose our king even deliberately to a bishop. I will add the same logic to rook's and queen's corresponding methods as well.