Observations & Fixes
We cannot check a king once we take an enemy pawn by en-passant. Also we were able to take an enemy queen which was guarded by an enemy bishop with our king, also look into resolving that too. I have fixed that we cannot take a guarded piece by using our king by alternating a function call in the take function of the piece class.