Observation & Fixes, game seems fine
I also observed that we can rook even if the path is moveable by a black enemy bishop positioned at g7 when we wanted to rook towards right. In addition, in an example game, I couldn't move the a1 rook of white king to any available squares for some reason. Also we cannot move the black king to the top right most square. In addition, we cannot take an unguarded piece with our king despite it's shown as a moveable square. Problem is prospectively stemming from the showMoveables of the king class. In fact I observed that showMoveaables has been working fine but invoking the isSquareThreatened function from take method of the piece class was made with wrong color.