Fixes made, !Check for future bishop problems
We have to add additional containers so that we can give user the ability to promote a pawn when applicable and also show additional information.  Furthermore, I am considering to add animations to PiecePanes when they are moved, if I have free time. There is an issue in bishop's showMoveables static version which is being invoked from isChecked.
!!!IMPORTANT:
If you face any problems with bishop revert changes made to bishop class in this commit.